Fix MileageController
