+++ build.gradle
@@ -31,6 +31,8 @@ dependencies {
 	compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.8.1'
 	compile group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
 
+	/*LocalDateTime 출력 관련*/
+	compile('net.sargue:java-time-jsptags:1.1.4')
 
 	compileOnly 'org.projectlombok:lombok'
 	compile('org.apache.tomcat.embed:tomcat-embed-jasper')
