--- etc/mypage-editMyInfo.jsp
@@ -1,41 +1,0 @@
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-12">
-			<h3>
-				회원정보수정
-			</h3>
-			<div class="row">
-				<div class="col-md-3">
-					<div class="btn-group btn-group-lg btn-group-vertical" role="group">
-						 
-						<button class="btn btn-secondary" type="button">
-							쇼핑정보
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							마일리지
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원정보수정
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원탈퇴
-						</button>
-					</div>
-				</div>
-				<div class="col-md-9">
-					<form role="form">
-						<div class="form-group">
-							<label for="exampleInputPassword1">
-								Password
-							</label>
-							<input type="password" class="form-control" id="exampleInputPassword1" />
-						</div>
-						<button type="submit" class="btn btn-warning">
-							확인
-						</button>
-					</form>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
