--- etc/mypage-editMyInfo2.jsp
@@ -1,126 +1,0 @@
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-12">
-			<h3>
-				회원정보수정
-			</h3>
-			<div class="row">
-				<div class="col-md-3">
-					<div class="btn-group btn-group-lg btn-group-vertical" role="group">
-						 
-						<button class="btn btn-secondary" type="button">
-							쇼핑정보
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							마일리지
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원정보수정
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원탈퇴
-						</button>
-					</div>
-				</div>
-				<div class="col-md-9">
-					<table class="table">
-						<thead>
-							<tr>
-								<th>
-									#
-								</th>
-								<th>
-									Product
-								</th>
-								<th>
-									Payment Taken
-								</th>
-								<th>
-									Status
-								</th>
-							</tr>
-						</thead>
-						<tbody>
-							<tr>
-								<td>
-									1
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									01/04/2012
-								</td>
-								<td>
-									Default
-								</td>
-							</tr>
-							<tr class="table-active">
-								<td>
-									1
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									01/04/2012
-								</td>
-								<td>
-									Approved
-								</td>
-							</tr>
-							<tr class="table-success">
-								<td>
-									2
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									02/04/2012
-								</td>
-								<td>
-									Declined
-								</td>
-							</tr>
-							<tr class="table-warning">
-								<td>
-									3
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									03/04/2012
-								</td>
-								<td>
-									Pending
-								</td>
-							</tr>
-							<tr class="table-danger">
-								<td>
-									4
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									04/04/2012
-								</td>
-								<td>
-									<button type="button" class="btn btn-warning">
-										배송지 목록
-									</button> 
-								</td>
-							</tr>
-						</tbody>
-					</table> 
-					
-					<button type="button" class="btn btn-warning btn-lg">
-						정보 수정
-					</button>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
