--- etc/mypage-myMileage.jsp
@@ -1,145 +1,0 @@
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-12">
-			<h3>
-				마일리지
-			</h3>
-			<div class="row">
-				<div class="col-md-3">
-					<div class="btn-group btn-group-lg btn-group-vertical" role="group">
-						 
-						<button class="btn btn-secondary" type="button">
-							쇼핑정보
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							마일리지
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원정보수정
-						</button> 
-						<button class="btn btn-secondary" type="button">
-							회원탈퇴
-						</button>
-					</div>
-				</div>
-				<div class="col-md-9">
-					<table class="table table-hover table-striped">
-						<thead>
-							<tr>
-								<th>
-									#
-								</th>
-								<th>
-									적립/사용금
-								</th>
-								<th>
-									적립/사용내역
-								</th>
-								<th>
-									적립/사용일
-								</th>
-							</tr>
-						</thead>
-						<tbody>
-							<tr>
-								<td>
-									1
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									01/04/2012
-								</td>
-								<td>
-									Default
-								</td>
-							</tr>
-							<tr class="table-active">
-								<td>
-									1
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									01/04/2012
-								</td>
-								<td>
-									Approved
-								</td>
-							</tr>
-							<tr class="table-success">
-								<td>
-									2
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									02/04/2012
-								</td>
-								<td>
-									Declined
-								</td>
-							</tr>
-							<tr class="table-warning">
-								<td>
-									3
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									03/04/2012
-								</td>
-								<td>
-									Pending
-								</td>
-							</tr>
-							<tr class="table-danger">
-								<td>
-									4
-								</td>
-								<td>
-									TB - Monthly
-								</td>
-								<td>
-									04/04/2012
-								</td>
-								<td>
-									Call in to confirm
-								</td>
-							</tr>
-						</tbody>
-					</table>
-				</div>
-			</div>
-			<nav>
-				<ul class="pagination">
-					<li class="page-item">
-						<a class="page-link" href="#">Previous</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">1</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">2</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">3</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">4</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">5</a>
-					</li>
-					<li class="page-item">
-						<a class="page-link" href="#">Next</a>
-					</li>
-				</ul>
-			</nav>
-		</div>
-	</div>
-</div>
