+++ src/main/java/com/pubble/conpub/domain/Member.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.io.Serializable;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,7 +15,7 @@ import java.util.List;
 @SequenceGenerator(
         name = "MEMBER_SEQ_GEN",
         sequenceName = "MEMBER_SEQ",
-        initialValue = 1000, //1000부터 일반회원
+        initialValue = 1001, //1001부터 일반회원
         allocationSize = 1
 )
 public class Member implements Serializable {
@@ -68,11 +70,6 @@ public class Member implements Serializable {
 
     @OneToMany(mappedBy = "smsMember")
     private List<SMS> sms = new ArrayList<SMS>();
-
-
-
-
-
 
 
     /*@Enumerated(EnumType.STRING)
