+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -1,6 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 		 pageEncoding="UTF-8"%>
-<%@include file ="./include/header.jsp" %>
+<%@include file ="include/header.jsp" %>
 
 		<!-- section -->
 			<div class="carousel slide" id="carousel-249579">
@@ -98,4 +98,4 @@
 					</dl>
 				</div>
 			</div>
-<%@include file ="./include/footer.jsp" %>
+<%@include file ="include/footer.jsp" %>
